Highlight Skyscanner when ready.
To get Skyscanner layover info, we need to open a modal for each flight.
We also need to close each of these modals, and this all takes time.
So if we have not finished closing these modals, do not take the user to that site. We will take them there once all modals have been closed.