Fix Expedia prematurely clearing return selections.
For roundtrip flights, if you pick a return, you are taken to Expedia, then go back to flightpenguin, and pick another return, then taken to Expedia again and see this return selected. The bug was that after returning back to flightpenguin, Expedia would go back to departures view because this was the original behavior, where returning to flightpenguin would clear all selections.