Fix highlighting Skyscanner itin for first time, more cases.
Skyscanner has a couple of running async tasks for parsing their HTML, so trying to highlight an itin while it is working is tricky if it is in the midst of a task, so add some checks within task to return early.
Also move showMoreResults call to prevent scroll to the bottom prior to highlight (not necessary).

Also scroll to the top prior to highlight itin.